ОХУЕННЫЙ такой фикс корзины (нет сущностей в DTO)

+++ src/main/java/marketplace/repository/basket/BasketRepo.java
@@ -21,7 +21,7 @@ public class BasketRepo {
         return basket;
     }
 
-    public Optional<Basket> find(int id) {
+    public Optional<Basket> find(long id) {
         return Optional.ofNullable(entityManager.find(Basket.class, id));
     }
